Changes and fixes
